refactor: Improve OAuth token route and user info route

+++ apps/web/app/api/oauth/token/route.ts
@@ -1,5 +1,5 @@
 import { DubApiError, handleAndReturnErrorResponse } from "@/lib/api/errors";
-import { TOKEN_EXPIRY, TOKEN_LENGTHS } from "@/lib/api/oauth";
+import { TOKEN_EXPIRY, TOKEN_LENGTH } from "@/lib/api/oauth";
 import { getAuthTokenOrThrow, hashToken } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import { codeExchangeSchema } from "@/lib/zod/schemas/oauth";
@@ -10,8 +10,12 @@ import { NextRequest, NextResponse } from "next/server";
 export async function POST(req: NextRequest) {
   try {
     const formData = Object.fromEntries(await req.formData());
-    const { client_id, client_secret, code, redirect_uri } =
-      codeExchangeSchema.parse(formData);
+    const {
+      client_id,
+      client_secret,
+      code,
+      redirect_uri: redirectUri,
+    } = codeExchangeSchema.parse(formData);
 
     let clientId = client_id;
     let clientSecret = client_secret;
@@ -37,8 +41,8 @@ export async function POST(req: NextRequest) {
     const oauthCode = await prisma.oAuthCode.findFirst({
       where: {
         code,
-        clientId: client_id,
-        redirectUri: redirect_uri,
+        clientId,
+        redirectUri,
         app: {
           clientSecretHashed: await hashToken(clientSecret),
         },
@@ -46,9 +50,10 @@ export async function POST(req: NextRequest) {
       select: {
         userId: true,
         projectId: true,
+        scopes: true,
+        expiresAt: true,
         app: {
           select: {
-            id: true,
             name: true,
           },
         },
@@ -62,36 +67,67 @@ export async function POST(req: NextRequest) {
       });
     }
 
-    await prisma.oAuthCode.delete({
+    if (oauthCode.expiresAt < new Date()) {
+      await prisma.oAuthCode.delete({
+        where: {
+          code,
+        },
+      });
+
+      throw new DubApiError({
+        code: "bad_request",
+        message: "Authorization code has expired.",
+      });
+    }
+
+    const workspace = await prisma.project.findUnique({
       where: {
-        code,
+        id: oauthCode.projectId,
+      },
+      select: {
+        plan: true,
       },
     });
 
-    const accessToken = `dub_${nanoid(TOKEN_LENGTHS.accessToken)}`;
-    const hashedKey = await hashToken(accessToken);
-    const partialKey = `${accessToken.slice(0, 3)}...${accessToken.slice(-4)}`;
+    const accessToken = `dub_${nanoid(TOKEN_LENGTH.accessToken)}`;
+    const expires = new Date(Date.now() + TOKEN_EXPIRY.accessToken);
+    const expiresInSeconds = Math.floor(expires.getTime() / 1000);
 
-    const { expires } = await prisma.restrictedToken.create({
-      data: {
-        name: `Access token for ${oauthCode.app.name}`,
-        hashedKey,
-        partialKey,
-        userId: oauthCode.userId,
-        projectId: oauthCode.projectId,
-        rateLimit: getCurrentPlan("free").limits.api,
-        expires: new Date(Date.now() + TOKEN_EXPIRY.accessToken),
-        scopes: "",
-        // scopes && scopes.length > 0 ? [...new Set(scopes)].join(" ") : null,
-      },
+    await Promise.all([
+      // Create the access token
+      prisma.restrictedToken.create({
+        data: {
+          name: `Access token for ${oauthCode.app.name}`,
+          hashedKey: await hashToken(accessToken),
+          partialKey: `${accessToken.slice(0, 3)}...${accessToken.slice(-4)}`,
+          userId: oauthCode.userId,
+          projectId: oauthCode.projectId,
+          scopes: oauthCode.scopes,
+          rateLimit: getCurrentPlan(workspace?.plan as string).limits.api,
+          expires: new Date(Date.now() + TOKEN_EXPIRY.accessToken),
+        },
+      }),
+
+      // Delete the code after it's been used
+      prisma.oAuthCode.delete({
+        where: {
+          code,
+        },
+      }),
+    ]);
+
+    console.log("Token exchange", {
+      access_token: accessToken,
+      refresh_token: accessToken,
+      token_type: "Bearer",
+      expires_in: expiresInSeconds,
     });
 
     return NextResponse.json({
       access_token: accessToken,
-      token_type: "Bearer",
-      expires_in: expires,
-      scope: "read write",
       refresh_token: accessToken,
+      token_type: "Bearer",
+      expires_in: expiresInSeconds,
     });
   } catch (error) {
     return handleAndReturnErrorResponse(error);
